device/requirements.txt
paho-mqtt==1.6.1
RPi.GPIO==0.7.1
dht11==0.1.0